requirements.txt
flask>=2.3.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
plotly>=5.15.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
openpyxl>=3.1.0
bcrypt>=4.0.0
